Change 'set' by 'store' to indicate that the info is store in the DataStore
