+++ app/src/main/java/ar/com/westsoft/listening/domain/dictationgame/settings/StoreReadWordAfterCursorUseCase.kt
@@ -4,17 +4,17 @@ import ar.com.westsoft.listening.data.datasource.PreferencesKey
 import ar.com.westsoft.listening.data.repository.SettingsRepository
 import javax.inject.Inject
 
-class SetSpeechRateUseCase @Inject constructor(
+class StoreReadWordAfterCursorUseCase @Inject constructor(
     private val settingsRepository: SettingsRepository
 ) {
     suspend operator fun invoke(value: String): Boolean {
-        val number = PreferencesKey.SpeechRatePercentage.convert(value) ?: return false
+        val number = PreferencesKey.ReadWordAfterCursor.convert(value) ?: return false
 
-        return if (PreferencesKey.SpeechRatePercentage.conditionToSave(number)) {
-            settingsRepository.setSpeechRate(number)
+        return if (PreferencesKey.ReadWordAfterCursor.conditionToSave(number)) {
+            settingsRepository.setReadWordAfterCursor(number)
         } else {
-            return false
+            false
         }
-
     }
 }
+
