+++ app/src/main/java/ar/com/westsoft/listening/domain/dictationgame/settings/StoreReadWordBeforeCursorUseCase.kt
@@ -4,16 +4,17 @@ import ar.com.westsoft.listening.data.datasource.PreferencesKey
 import ar.com.westsoft.listening.data.repository.SettingsRepository
 import javax.inject.Inject
 
-class SetColumnPerPageUseCase @Inject constructor(
+class StoreReadWordBeforeCursorUseCase @Inject constructor(
     private val settingsRepository: SettingsRepository
 ) {
     suspend operator fun invoke(value: String): Boolean {
-        val number = PreferencesKey.ColumnPerPage.convert(value) ?: return false
+        val number = PreferencesKey.ReadWordBeforeCursor.convert(value) ?: return false
 
-        return if (PreferencesKey.ColumnPerPage.conditionToSave(number)) {
-            settingsRepository.setColumnPerPage(number)
+        return if (PreferencesKey.ReadWordBeforeCursor.conditionToSave(number)) {
+            settingsRepository.setReadWordBeforeCursor(number)
         } else {
             false
         }
+
     }
 }
